DAT-565 Integration test for loading Countries

+++ fd-client/src/main/java/org/flockdata/client/Importer.java
@@ -116,7 +116,7 @@ public class Importer extends SpringBootServletInitializer {
         }
 
         StopWatch watch = new StopWatch("Batch Import");
-        long totalRows = 0;
+        int totalRows = 0;
         try {
 
             int batchSize = clientConfiguration.getBatchSize();
